Run checks every minute in the background.
Whenever things are busy/broken everyone (understandably) jumps onto the
aggregate healthcheck to see whats wrong.  For each request, we kick of
n*more requests to the back end apps.  The end result contributes to
massive network usage, and we've seen tens of thousands of tcp
connections in the agg hc docker container.

This change means we update a stored view of service health every
minute, and when requesting we always return the same cached result
immediately.